+++ MordoEngine/src/Scene/Game/GameScene.h
@@ -4,7 +4,7 @@
 #include "../../Terrain/Terrain.h"
 #include "../../Renderer/Renderer.h"
 #include "../../Lighting/Sun.h"
-#include "../Editor/Controllers/EditorCameraController.h"
+#include "./Controllers/GameCameraController.h"
 #include <memory>
 
 class GameScene : public Scene
@@ -13,12 +13,13 @@ private:
 	std::shared_ptr<terrain::Terrain> m_Terrain;
 	std::shared_ptr<Camera> m_Camera;
 	std::shared_ptr<Renderer> m_Renderer;
-	std::unique_ptr<EditorCameraController> m_CameraController;
+	std::unique_ptr<GameCameraController> m_CameraController;
 	std::unique_ptr<Sun> m_Sun;
 
 public:
 	GameScene(std::shared_ptr<terrain::Terrain> terrain, std::shared_ptr<Camera> camera, std::shared_ptr<Renderer> renderer);
 	void Render() override;
 	void Update(float deltaTime) override;
+	void OnEntry() override;
 	~GameScene() override;
 };