+++ MordoEngine/src/Scene/Scene.h
@@ -12,5 +12,6 @@ protected:
 public:
 	virtual void Update(float deltaTime) = 0;
 	virtual void Render() = 0;
+	virtual void OnEntry() = 0;
 	virtual ~Scene() = default;
 };